Added base class for user-defined scenarios
User-defined scenarios shall now be derived from BasicScenario defined
in basic_scenario.py
